fixed by review notes

+++ hudi-spark-datasource/hudi-spark-common/src/main/scala/org/apache/spark/sql/execution/datasources/lance/SparkLanceReaderBase.scala
@@ -86,8 +86,8 @@ class SparkLanceReaderBase(enableVectorizedReader: Boolean) extends SparkColumna
       try {
         val lanceReader = LanceFileReader.open(file.filePath.toString, countAllocator)
         try {
-          val rowCount = lanceReader.numRows()
-          Iterator.fill(rowCount.toInt)(InternalRow.empty)
+          val rowCount = Math.toIntExact(lanceReader.numRows())
+          Iterator.fill(rowCount)(InternalRow.empty)
         } finally {
           lanceReader.close()
         }
